Split src/asm/charmap.c into two files
This way it is easier to identify the license of the code.
