+++ include/extern/utf8decoder.h
@@ -1,0 +1,6 @@
+#ifndef EXTERN_UTF8DECODER_H
+#define EXTERN_UTF8DECODER_H
+
+uint32_t decode(uint32_t *state, uint32_t *codep, uint32_t byte);
+
+#endif /* EXTERN_UTF8DECODER_H */
